+++ user/src/main/java/org/abr/audreybr/dto/ChouilleDTO.java
@@ -14,21 +14,21 @@ import java.sql.Date;
 @Setter
 public class ChouilleDTO implements Serializable {
 
-    private Long Id_Chouille;
+    private Integer Id_Chouille;
 
     private String Thematic;
 
     private Date Date;
 
-    private Long Id_Location;
+    private Integer Id_Location;
 
-    private Long Id_Person_Sam;
+    private Integer Id_Person_Sam;
 
-    private Long Id_Person_Bouncer;
+    private Integer Id_Person_Bouncer;
 
     private String code;
 
-    public ChouilleDTO(Long id_Chouille, String thematic, java.sql.Date date, Long id_Location, Long id_Person_Sam, Long id_Person_Bouncer, String code) {
+    public ChouilleDTO(Integer id_Chouille, String thematic, java.sql.Date date, Integer id_Location, Integer id_Person_Sam, Integer id_Person_Bouncer, String code) {
         Id_Chouille = id_Chouille;
         Thematic = thematic;
         Date = date;
@@ -38,11 +38,11 @@ public class ChouilleDTO implements Serializable {
         this.code = code;
     }
 
-    public Long getId_Chouille() {
+    public Integer getId_Chouille() {
         return Id_Chouille;
     }
 
-    public void setId_Chouille(Long id_Chouille) {
+    public void setId_Chouille(Integer id_Chouille) {
         Id_Chouille = id_Chouille;
     }
 
@@ -62,27 +62,27 @@ public class ChouilleDTO implements Serializable {
         Date = date;
     }
 
-    public Long getId_Location() {
+    public Integer getId_Location() {
         return Id_Location;
     }
 
-    public void setId_Location(Long id_Location) {
+    public void setId_Location(Integer id_Location) {
         Id_Location = id_Location;
     }
 
-    public Long getId_Person_Sam() {
+    public Integer getId_Person_Sam() {
         return Id_Person_Sam;
     }
 
-    public void setId_Person_Sam(Long id_Person_Sam) {
+    public void setId_Person_Sam(Integer id_Person_Sam) {
         Id_Person_Sam = id_Person_Sam;
     }
 
-    public Long getId_Person_Bouncer() {
+    public Integer getId_Person_Bouncer() {
         return Id_Person_Bouncer;
     }
 
-    public void setId_Person_Bouncer(Long id_Person_Bouncer) {
+    public void setId_Person_Bouncer(Integer id_Person_Bouncer) {
         Id_Person_Bouncer = id_Person_Bouncer;
     }
